Bug fixes to allow multiple instances of a timepicker to exist, even in nested loops
