+++ src/react/components/Popover.tsx
@@ -14,7 +14,6 @@ import {
 } from "react";
 import { formatTime } from "../hooks/timeUtils";
 import { useClickOutside } from "../hooks/useClickOutside";
-
 import {
   PopoverContentProps,
   PopoverContextType,
